+++ @xen-orchestra/web-core/lib/locales/en.json
@@ -32,11 +32,11 @@
   "action:delete": "Delete",
   "action:delete-filter": "Delete filter",
   "action:delete-group": "Delete group",
+  "action:delete-n-vbds": "Delete 1 VBD | Delete {n} VBDs",
   "action:delete-n-vdis": "Delete 1 VDI | Delete {n} VDIs",
   "action:delete-n-vifs": "Delete 1 VIF | Delete {n} VIFs",
   "action:delete-n-vms": "Delete 1 VM | Delete {n} VMs",
-  "action:detach-n-vdis": "Detach 1 VDI | Detach {n} VDIs",
-  "action:detach": "Detach",
+  "action:delete-vbd": "Delete VBD",
   "action:deploy": "Deploy",
   "action:deploy-xoa": "Deploy XOA",
   "action:download-bugtools-archive": "Download bugtools archive",
@@ -383,9 +383,9 @@
   "job:arg:vlan-required": "VLAN is required",
   "job:create:in-progress": "Create in progress…",
   "job:delete:in-progress": "Delete in progress…",
-  "job:vdi-delete:missing-vdi": "No VDIs to delete",
   "job:vbd-delete:missing-vbd": "No VBDs to delete",
-  "job:vbd-delete:vbd-attached": "Cannot detach a VDI that is currently attached",
+  "job:vbd-delete:vbd-attached": "Cannot delete a VBD that is currently attached",
+  "job:vdi-delete:missing-vdi": "No VDIs to delete",
   "job:vdi-delete:vbd-attached": "Cannot delete a VDI with VBD attached",
   "job:vif-delete:blocked-operation": "VIF deletion is blocked",
   "job:vif-delete:missing-vif": "No VIFs to delete",
@@ -907,12 +907,12 @@
   "user-config": "User config",
   "username": "Username",
   "uuid": "UUID",
+  "vbd-delete-info": "This will detach the selected VDI from the VM. The disk will not be deleted and can be reattached later.",
+  "vbd-delete-title": "You are about to delete 1 VBD | You are about to delete {n} VBDs",
   "vcpu": "vCPU",
   "vcpus": "vCPUs",
   "vcpus-assigned": "vCPUs assigned",
   "vdi": "VDI",
-  "vdi-detach-info": "This will detach the selected VDI from the VM. The disk will not be deleted and can be reattached later.",
-  "vdi-detach-title": "You are about to detach 1 VDI | You are about to detach {n} VDIs",
   "vdi-delete-warning": "This will permanently delete the selected VDI and all data stored on it. This action cannot be undone.",
   "vdi-throughput": "VDI throughput",
   "vdis": "VDIs",
