+++ @xen-orchestra/web-core/lib/locales/fr.json
@@ -32,11 +32,11 @@
   "action:delete": "Supprimer",
   "action:delete-filter": "Supprimer le filtre",
   "action:delete-group": "Supprimer le groupe",
-  "action:detach": "Détacher",
-  "action:detach-n-vdis": "Détacher 1 VDI | Détacher {n} VDIs",
+  "action:delete-n-vbds": "Supprimer 1 VBD | Supprimer {n} VBDs",
   "action:delete-n-vdis": "Supprimer 1 VDI | Supprimer {n} VDIs",
   "action:delete-n-vifs": "Supprimer 1 VIF | Supprimer {n} VIFs",
   "action:delete-n-vms": "Supprimer 1 VM | Supprimer {n} VMs",
+  "action:delete-vbd": "Supprimer le VBD",
   "action:deploy": "Déployer",
   "action:deploy-xoa": "Déployer XOA",
   "action:download-bugtools-archive": "Télécharger l’archive bugtools",
@@ -380,14 +380,11 @@
   "job:arg:sr-vdi-required": "Le VDI doit avoir un SR",
   "job:arg:template-uuid-required": "L'UUID de template est obligatoire",
   "job:arg:username-required": "Le nom d'utilisateur est obligatoire",
-  "job:delete:in-progress": "Suppression en cours…",
-  "job:vdi-delete:missing-vdi": "Aucun VDI à supprimer",
-  "job:vbd-delete:missing-vbd": "Aucun VBD à supprimer",
-  "job:vbd-delete:vbd-attached": "Impossible de détacher un VDI actuellement attaché",
-  "job:vdi-delete:vbd-attached": "Impossible de supprimer un VDI avec un VBD attaché",
   "job:arg:vlan-required": "Le VLAN est obligatoire",
   "job:create:in-progress": "Création en cours…",
   "job:delete:in-progress": "Suppression en cours…",
+  "job:vbd-delete:missing-vbd": "Aucun VBD à supprimer",
+  "job:vbd-delete:vbd-attached": "Impossible de supprimer un VBD actuellement attaché",
   "job:vdi-delete:missing-vdi": "Aucun VDI à supprimer",
   "job:vdi-delete:vbd-attached": "Impossible de supprimer un VDI avec un VBD attaché",
   "job:vif-delete:blocked-operation": "La suppression du VIF est bloquée",
@@ -910,12 +907,12 @@
   "user-config": "Configuration utilisateur",
   "username": "Nom d'utilisateur",
   "uuid": "UUID",
+  "vbd-delete-info": "Cela détachera le VDI sélectionné de la VM. Le disque ne sera pas supprimé et pourra être rattaché ultérieurement.",
+  "vbd-delete-title": "Vous êtes sur le point de supprimer 1 VBD | Vous êtes sur le point de supprimer {n} VBDs",
   "vcpu": "vCPU",
   "vcpus": "vCPUs",
   "vcpus-assigned": "vCPUs assignés",
   "vdi": "VDI",
-  "vdi-detach-info": "Cela détachera le VDI sélectionné de la VM. Le disque ne sera pas supprimé et pourra être rattaché ultérieurement.",
-  "vdi-detach-title": "Vous êtes sur le point de détacher 1 VDI | Vous êtes sur le point de détacher {n} VDIs",
   "vdi-delete-warning": "Cela supprimera définitivement le VDI sélectionné ainsi que toutes les données qu'il contient. Cette action est irréversible.",
   "vdi-throughput": "Débit du VDI",
   "vdis": "VDIs",
